+++ public/data/promotionData.json
@@ -1,16 +1,16 @@
 [
   {
     "id": 1,
-    "quiz": "퀴즈 프로모션 페이지 - 아이유의 별명은 이지금이다."
+    "quizImg": "./images/quizImg.jpg"
   },
 
   {
     "id": 2,
-    "quiz": "퀴즈 프로모션 페이지 - 에픽하이는 정규 10집까지 발매했다."
+    "quizImg": "퀴즈 프로모션 페이지 - 에픽하이는 정규 10집까지 발매했다."
   },
 
   {
     "id": 3,
-    "quiz": "퀴즈 프로모션 페이지 - 거미는 올해 상반기에 전국투어를 진행했다."
+    "quizImg": "퀴즈 프로모션 페이지 - 거미는 올해 상반기에 전국투어를 진행했다."
   }
 ]
